reorganize component folder structure
