+++ src/components/FormNumberField/index.ts
@@ -1,0 +1,2 @@
+import component from './FormNumberField';
+export default component;
